requirements.txt
beautifulsoup4==4.9.3
blis==0.7.4
bs4==0.0.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cymem==2.0.5
elementpath==2.2.1
idna==2.10
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
meteo-downloader==1.0
murmurhash==1.0.5
nltk==3.5
numpy==1.20.2
packaging==20.9
parse==1.19.0
pathy==0.4.0
pdfminer.six==20201018
preshed==3.0.5
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.2.1
srsly==2.4.0
thinc==8.0.2
tqdm==4.59.0
typer==0.3.2
urllib3==1.26.4
wasabi==0.8.2
xmlschema==1.6.1
